chore: Move shared to lib/ for AWS
