+++ providers/aws/lib/arn/arn_test.py
@@ -1,6 +1,6 @@
 import sure  # noqa
 
-from lib.arn.arn import arn_parsing
+from providers.aws.lib.arn.arn import arn_parsing
 
 ACCOUNT_ID = "123456789012"
 RESOURCE_TYPE = "role"
