+++ providers/aws/lib/arn/error.py
